tools/article_editor/requirements.txt
deep-translator==1.11.4
opencc-python-reimplemented==0.1.7
pyinstaller==6.16.0
